Move definition of std::hash<DomineeringState> to DomineeringState.h
This change is made so that std::hash<DomineeringState> can be used in future
implementations. Without this change, those implementations need to include
Searcher.h, which is not appropriate and may result in circular dependencies.
