+++ src/common/DomineeringState.h
@@ -34,5 +34,28 @@ private:
 
 };
 
+/**
+ * Combine operation of two hash keys. Based on boost::hash_combine.
+ */
+namespace {
+    void hash_combine(std::size_t& h, const std::size_t& v) {
+        h ^= v + 0x9e3779b9 + (h << 6) + (h >> 2);
+    }
+}
+
+namespace std {
+    template<>
+    struct hash<DomineeringState> {
+        size_t operator()(const DomineeringState& ds) const {
+            size_t h = 0;
+            for (auto&& c : *ds.getBoard1D()) {
+                if (c == ds.EMPTYSYM) {
+                    hash_combine(h, std::hash<char>()(c));
+                }
+            }
+            return h;
+        }
+    };
+} // namespace std
 
 #endif /* defined(__CSE486AIProject__DomineeringState__) */
